requirements.txt
reflex>=0.3.6
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
bcrypt>=4.0.1
python-jose[cryptography]>=3.3.0
pytest>=7.4.0
stripe>=6.0.0
pillow>=10.0.0
python-multipart>=0.0.6
email-validator>=2.0.0
